requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.1
deprecation==2.1.0
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.43.1
frozenlist==1.4.0
gotrue==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.301
langsmith==0.0.40
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pinecone-client==2.2.4
plotly==5.17.0
postgrest==0.10.8
pydantic==2.4.0
pydantic_core==2.10.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.0
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
websockets==10.4
Werkzeug==2.3.7
yarl==1.9.2
